+++ Sources/FLCharts/Models/Configurations/FLLineConfig.swift
@@ -22,14 +22,18 @@ public struct FLLineConfig {
     /// Whether to show a circle that represents a specific point.
     public var showCircles: Bool
     
+    /// Whether to show a smooth quadratic line or a segmented one.
+    public var isSmooth: Bool
+
     /// The configuration of the circles.
     public var circleColor: UIColor
     
-    public init(width: CGFloat = 4, capStyle: CGLineCap = .round, backgroundFill: FLColor = FLColor.Gradient.lightBlue, showCircles: Bool = false, circleColor: UIColor = .white) {
+    public init(width: CGFloat = 4, capStyle: CGLineCap = .round, backgroundFill: FLColor = FLColor.Gradient.lightBlue, showCircles: Bool = false, isSmooth: Bool = true, circleColor: UIColor = .white) {
         self.width = width
         self.capStyle = capStyle
         self.showCircles = showCircles
         self.circleColor = circleColor
+        self.isSmooth = isSmooth
         self.backgroundFill = backgroundFill
     }
 }